extra alerts state into its own module in store
