+++ packages/client/src/store/index.js
@@ -8,14 +8,11 @@ import agencies from './modules/agencies';
 import dashboard from './modules/dashboard';
 import organization from './modules/organization';
 import tenants from './modules/tenants';
+import alerts from './modules/alerts';
 
 Vue.use(Vuex);
 
 const debug = process.env.NODE_ENV !== 'production';
-
-function randomId() {
-  return Math.random().toString(16).substr(2, 10);
-}
 
 export default new Vuex.Store({
   strict: debug,
@@ -27,16 +24,6 @@ export default new Vuex.Store({
     dashboard,
     organization,
     tenants,
-  },
-  state: {
-    alerts: {},
-  },
-  mutations: {
-    addAlert(state, alert) {
-      Vue.set(state.alerts, randomId(), alert);
-    },
-    dismissAlert(state, alertId) {
-      Vue.delete(state.alerts, alertId);
-    },
+    alerts,
   },
 });
