+++ packages/client/src/store/modules/users.js
@@ -7,7 +7,6 @@ function initialState() {
       selectedAgencyId: null,
     },
     users: [],
-    alerts: {},
   };
 }
